api/backend/requirements.txt
flask
flask-cors
gunicorn
numpy
google-cloud-aiplatform
